[Feature] Redesigned return instruction
It used to be a epilogue at the end of the function, but now we provide
transformation for each return keyword. This considers two factors
1) L2 is more complex than L1, because there may exist multiple return
in different branches. The old logic is once we face a return, we do not
continue generate following IR. The new logic will keep transforming
after read a Return.
2) function epilogue is in some way executed at the same time when
return executed. So it is quite normal to embed it to a instruction,
instead of string.