[Bugfix] Fixed a bug for register assignment of declared variable
Some variables may only be declared and not defined in the whole
program. If it used on the RHS of exp, then it is typechecked. However,
if it is used in Return, previous program cannot map a register to it.

Now we use %eax to map those only declared but not defined temporaries.
Since those temporaries are only declared ant not assigned, so we don't
need to worry that %eax may be modified.