[Bugfix] Provide L2 context const propagation
Notice that copy propagation is also modified. But it does not work yet
because when we process AST -> IR, we cannot assign a new temporary for
each variable like int a = 3; a = a + 1; because we need to consider if
statement. We can easily handle it before as
```A.Assign (id, e) :: stms ->
   let t = Temp.create () in
   let env' = S.set env ~key:id ~data:t in
```
But we cannot easily handle env in if statement. So we break SSA for now
and do not optimize copy propagation yet.