[Bugfix] Sexp is corrected.
IR Sexp is an expression, so it is valid for statement like

a = (b : Sexp)

So in this context, Sexp should be returned to a. It is wrong to say
that Sexp is used only for side effect and does not any return value.

Sexp can be generated from ternary operator.