[Optimize] Optimized x86 cmp code
We can use eax without store its value at the beginning of transcode
from pseudo to x86 assembly.

Remember that eax is preserved during register allocation. So it will
not store any temporary. Though it may be used within div, mul, but once
thoses instruction are finished, the dest are stored in corresponding
regisetrs. So eax and other preserved registers can be used for free.