[Feature] Distinguish exp with side-effect in IR
Now expression in IR.exp is just pure expression.

Expression with side effect is only reside in IR.stm

This framework can extend to function call in lab3.

This makes SSA calculating def use more cohesive. The compiler take Sexp
as a whole in the before, and hacked in this process. Now label can only
appear in statement, and will not be part of a sexp of exp type.