[Feature] Use label as map index in dom and ssa
So each block map in dominator and ssa has key of label. Also, its
successors are label set now. This framework makes block more flexible
compared with previous using int as block number.

[Bugfix]

Fixed a bug in finding least common subroot across a node during idom
procedure. Ths first new_idom should be a PROCESSED parent of node n. In
other words, if idom_map has a node n, it means this node and its
parents(at least part of) are processes. And during intersect function,
we don't need to warry that idom_map[b1] returns None.

To make it simple, if idom_map[n] is not None, it means there is a path
from entry to n in the graph.