[Feature] Provide frame memory
Frame memory is used to pass parameters across function call. It is
treated as a immediate in def/use/liveout because it can only be read.

Abs_asm can resort this operand type and don't bother much in regalloc.
It will be translated to memory access in x86_asm codegen.