[Optimize] Save 50% stack memory for temporary
Compiler used 8 bytes to store temporary, which is a waste of space
because operand is of 4 bytes. Now use operand size to allocate memory,
which is more flexible and efficient.

Notice that now there is only immediate, so there is no overlap check.
It is possible that 1th memory and 2nd memory has overlap if they are
struct. Allocation with overlap check will be done in the future.