[Feature] Handle function with void return
The compiler used to generate a dummy temp as destination for void
function. However, when we provide size for each temp, this temp has
size of void, and can lead to be assigned a register with size void,
which is non-sense. New feature now

1) IR does not create dummy temp. Instead, IR.exp has new type of void.
2) QUADS does not generate move void -> temp
3) abs_asm fcall still preserve %rax

[Feature] Distinguish logic register and hard register

%rax, %eax, %ax, %al is the same logic register, but they are different
hard register(though they are on the same real register).

1) abs_asm use hard register to indicate which part of register should
be preserved, considering the size of parameters, return, and
caller/calee registers.
2) regalloc use logic register as IG.Vertex. So no worry for too much
vertex for the same logic register.
3) logic register is not assigned in regalloc, instead, they are
assigned in x86_asm. The size for each register is based on the
information stored in line and define.