[Feature] L4 lexer, parser and CST
Though the grammar has lvalue, we do not provide rule in parser because
lvalue and exp will lead to reduce-reduce conflict in simp when parsing
identifier.

Instead, lvalue is replaced with exp. This works because lvalue is a
subset of exp. This will make some source code that does not fit L4
grammar parse, but it will be checked in later elaboration pass.