[Feature] AST->IR framework
Provide new type for expression and lvalue, edest and ldest
respectively.

When array, struct and dereference is allowed, value is not only
described by expression, but also memory. So the return during
transformation from AST exp -> IR exp now have memory.

This type is only a intermediate representation during transformation,
and will be eliminated when transformation done. For memory, it will be
transformed to load and store statement in IR.