[Feature] Use %rbp as base pointer
I was trying to remove %rbp as base pointer, and use it as index pointer

However, it would be very hard to implement push if we just use %rsp to

track all spilled variables and pushed variables because %rsp will change

during push and pop.

Therefore, I reimport %rbp as stack base pointer.